requirements.txt
imutils
gunicorn
Django==1.9
cmake
djangorestframework==3.6.3
dlib==19.17.0
face_recognition==1.2.3
requests
opencv-python-headless
